Memperbaiki tampilan order dan costumer

+++ database/migrations/2025_01_29_184510_create_customers_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string("customer_id")->unique();
             $table->string("Customer_Name");
-            $table->integer("Phone Number");
+            $table->bigInteger("Phone Number");
             $table->string("Email");
             $table->string("Occupation");
             $table->date("Date of Birth");
